refactor: auth structure
