+++ src/features/auth/index.ts
@@ -1,7 +1,2 @@
-export { default as AuthForm } from './ui/AuthForm';
-export { default as PasswordVisibility } from './ui/PasswordVisibility';
-export {
-  RegistrationFormSchema,
-  type RegistrationInputType,
-} from './model/schemas/registrationFormSchema';
-export { LoginFormSchema, type LoginInputType } from './model/schemas/loginFormSchema';
+export { default as LoginForm } from './ui/LoginForm';
+export { default as RegistrationForm } from './ui/RegistrationForm';
