--- src/test.ts
@@ -1,1 +1,0 @@
-export let theme = 'light';
