+++ src/views/registration/ui/RegistrationPage.tsx
@@ -1,5 +1,5 @@
-import { RegistrationForm } from '@/widgets/registration';
 import { AuthLayout } from '@/widgets/auth-layout';
+import { RegistrationForm } from '@/features/auth';
 import * as React from 'react';
 
 const RegistrationPage: React.FC = () => {
