--- src/widgets/login/index.ts
@@ -1,1 +1,0 @@
-export { default as LoginForm } from './ui/LoginForm';
