--- src/widgets/registration/index.ts
@@ -1,1 +1,0 @@
-export { default as RegistrationForm } from './ui/RegistrationForm';
